requirements.txt
google-api-python-client==2.97.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.1.0
openai==1.31.0
python-dotenv==1.0.0
pandas==2.3.0
html2text==2020.1.16
pytest==8.0.0
pytest-mock==3.14.0
pytest-cov==4.1.0
